client-qna, server-answer, inquiry 수정
