+++ server/src/main/java/com/avengers/yoribogo/answer/domain/Answer.java
@@ -1,6 +1,7 @@
 package com.avengers.yoribogo.answer.domain;
 
 import com.avengers.yoribogo.common.Role;
+import com.avengers.yoribogo.user.domain.UserEntity;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -27,16 +28,17 @@ public class Answer {
     @Column(name = "answer_created_at")
     private LocalDateTime answerCreatedAt;
 
-    @Column(name = "user_id")
-    private int userId;
-
     @Column(name = "inquiry_id")
     private int inquiryId;
 
-    public Answer(String answerContent, Role writerType, int userId, int inquiryId) {
+    @OneToOne
+    @JoinColumn(name = "user_id", referencedColumnName = "user_id")
+    private UserEntity user;
+
+    public Answer(String answerContent, Role writerType,int inquiryId, UserEntity user) {
         this.answerContent = answerContent;
         this.writerType = writerType;
-        this.userId = userId;
         this.inquiryId = inquiryId;
+        this.user = user;
     }
 }
